Pulling customization from DIDs
