+++ cardstack/src/graphql/fragments.graphql
@@ -29,6 +29,7 @@ fragment PrepaidCardPayment on PrepaidCardPayment {
   }
   prepaidCard {
     id
+    customizationDID
   }
   merchantSafe {
     id
